spring security method security

+++ src/main/java/arda/xmlproject/demo/controller/UserController.java
@@ -5,13 +5,18 @@ import arda.xmlproject.demo.entities.UserEntity;
 import arda.xmlproject.demo.entities.UserRoles;
 import arda.xmlproject.demo.mappers.Mapper;
 import arda.xmlproject.demo.services.UserService;
+import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import static arda.xmlproject.demo.controller.AuthController.hashPassword;
 
 @RestController
 @RequestMapping("/api")
@@ -26,17 +31,20 @@ public class UserController {
     }
 
     @PostMapping(path = "/users")
-    public UserDto createUser(@RequestBody UserDto user) {
+    public UserDto createUser(@Valid @RequestBody UserDto user) {
         UserEntity userEntity = userMapper.mapFrom(user);
 
         userEntity.setRole(UserRoles.USER);
         userEntity.setPermissions(null);
+
+        userEntity.setPassword(hashPassword(userEntity.getPassword()));
 
         UserEntity savedUserEntity = userService.save(userEntity);
         return userMapper.mapTo(savedUserEntity);
 
     }
 
+    @PreAuthorize("hasAnyAuthority('admin', 'user_admin')")
     @GetMapping(path = "/users")
     public List<UserDto> listUsers() {
         List<UserEntity> users = userService.findAll();
@@ -45,6 +53,7 @@ public class UserController {
                 .collect(Collectors.toList());
     }
 
+    @PreAuthorize("hasAuthority('admin')")
     @GetMapping(path = "/users/{id}")
     public ResponseEntity<UserDto> getUser(@PathVariable("id") Long id) {
         Optional<UserEntity> foundUser = userService.findOne(id);
@@ -55,18 +64,24 @@ public class UserController {
         }).orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
+    @PreAuthorize("hasAnyAuthority('admin', 'user_admin')")
     @PutMapping(path = "/users/{id}")
     public ResponseEntity<UserDto> fullUpdateUser(@PathVariable("id") Long id, @RequestBody UserDto userDto) {
         if (!userService.isExists(id)) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity currentUser = userService.findByUsername(currentUsername).orElseThrow();
+
+        if (currentUser.getId().equals(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         userDto.setId(id);
         UserEntity userEntity = userMapper.mapFrom(userDto);
 
-        userEntity.setRole(null);
-        userEntity.setPermissions(null);
-
+        userEntity.setPassword(hashPassword(userDto.getPassword()));
 
         UserEntity savedUserEntity = userService.save(userEntity);
 
@@ -74,15 +89,19 @@ public class UserController {
     }
 
 
+    @PreAuthorize("hasAnyAuthority('admin', 'user_admin')")
     @PatchMapping(path = "/users/{id}")
     public ResponseEntity<UserDto> partialUpdate(@PathVariable("id") Long id, @RequestBody UserDto userDto) {
         if (!userService.isExists(id)) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
-        userDto.setRole(null);
-        userDto.setPermissions(null);
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity currentUser = userService.findByUsername(currentUsername).orElseThrow();
 
+        if (currentUser.getId().equals(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         UserEntity userEntity = userMapper.mapFrom(userDto);
         UserEntity updatedUser = userService.partialUpdate(id, userEntity);
@@ -91,10 +110,18 @@ public class UserController {
                 HttpStatus.OK);
     }
 
+    @PreAuthorize("hasAnyAuthority('admin', 'user_admin')")
     @DeleteMapping(path = "/users/{id}")
-    public ResponseEntity deleteUser(@PathVariable("id") Long id) {
+    public ResponseEntity<Void> deleteUser(@PathVariable("id") Long id) {
+
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity currentUser = userService.findByUsername(currentUsername).orElseThrow();
+
+        if (currentUser.getId().equals(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         userService.delete(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
-
 }
